@@ -1,2 +1,2 @@
 /* eslint-disable import/prefer-default-export */
-export { default as VuseSample } from './vuse-sample.vue';
+export { default as vuseButton } from './vuseButton';
